42.1. Filtering Products{Listen to search-products Event in Catalog}
Ok: we have products and searchTerm data. Let's update the searchTerm when the search-bar component tells us it's changed.

Listen to search-products Event in Catalog
Remember, in that component, we're dispatching an event called search-products. In Catalog, up in the template, find the <search-bar element and add @search-products="onSearchProducts", which is a method that we now need to create.
Down in the code, do it: add methods: {} and then onSearchProducts(). Since this method is going to be called when an event is emitted, it will receive an event argument. And because, when we emitted the event, we added a term key, we can use that here! We can say this.searchTerm = event.term.
Check it in Vue Dev Tools
Let's go check it out! Back on the dev tools... if I type disc, the searchTerm in SearchBar, of course, updates. But if we look in Catalog, the search term also changed here! Yes!