43.1. Async Computed Properties{Adding Search to the Product Service; AJAX inside a Computed Property?}
Our search is cool and fast... but it won't work if we start paginating the products... and the logic used in the search is super basic: we're just comparing the product name. In most cases, I'll want my search to be smarter than this: like via a more complex database query or by using something like Elasticsearch.
For both of these reasons, I want to refactor our search to fetch the data from the server, instead of doing it all in JavaScript.
Open a new tab and go to /api/products.jsonld: our shortcut to see what our API looks like. I've already added a basic filter to the API: you can add ?name= and a term, and it will filter the results. This actually only filters on the name property. But you can make the search in your API as rich as you need. The important part is that our API has a way for us to send a search query and get a sub-set of results.

Adding Search to the Product Service
Back in our code, open up services/products-service.js. This function is responsible for making the AJAX request for products and it can already filter by category. Now, add a second argument called searchTerm. Then, very simply, if we have a searchTerm, say params.name = searchTerm to add the query parameter. I'll even document the new param: it will be a string or null and it's called searchTerm.
Very nice!

AJAX inside a Computed Property?
Let's use this in catalog.vue. I'm going to do this in the simplest way possible first... and it's totally not going to work! We currently have a filteredProducts computed property, which computes the correct array of Products to use based on the searchTerm and products.
So the easiest thing to do is: if there is no searchTerm, return this.products. Else, if there is a searchTerm, let's make our API call! I'll copy the fetchProducts() line from below and say const response = await fetchProducts(). And, of course, Webpack is mad because this method now needs to be async.
Much better! To finish the function, I'll go steal some more code and say return response.data['hydra:member'].
So... this makes sense, right? When we reference filteredProducts in the template, that will call our function, we make the AJAX call, wait for it to finish and then return the new array of products. Genius!
But... you can already see that we have an angry underline below await. Bah... let's ignore that and try it! I'll refresh and... oh... we are broken. And we have an error:

 Invalid prop: type check failed for prop "products". Expected Array got Promise.

This is coming from ProductList and that prop is passed by Catalog. Yep, when we render <product-list, the filteredProducts is no longer an Array. It's a Promise!
And... yea. That makes sense. When you make a function async, when someone calls that function, it finishes immediately and returns a Promise: it does not actually wait for your code to run. Then, later, when your function does finally finish its work, that promise resolves.
So this is a long way of saying that computed properties can only do synchronous work: they must return a value immediately. If you try to do something async, then you'll end up returning a Promise instead of the real value.
Nope, if you need to do something asynchronous, then you can't use a computed property. So, computed properties are "yay! " for calculating synchronous stuff but are "boo! " for async stuff.