43.2. Async Computed Properties{Adding a new Data; Updating products on searchTerm Change}
Adding a new Data
The solution is to abandon your project and take up a peaceful career herding sheep. Or, you'll need to convert the computed property into a piece of data. And then, whenever that data needs to change - like whenever the searchTerm changes - you'll call a method that will make the AJAX call and update that data once its done.
So... let's do this! The first step is to add a piece of data that can hold the filtered products. But, actually, now that we're not trying to keep all the products in the products data so that we can filter based off of it, whenever our AJAX call finishes, it's now ok to change that products array directly. So instead of adding a new piece of data - we'll just change products.
Sweet! This means that, up in the template, we should change filteredProducts back to products. And back down, we can remove the computed section entirely.

Updating products on searchTerm Change
Here's the plan then: whenever the searchTerm changes, we basically want to re-run all of the code that makes the AJAX call and updates the products data... but with a minor addition to also include the search query.
To help re-use this, create a new method called loadProducts() with a searchTerm argument.
Now, copy the entire created() function... and paste. To include the searchTerm, pass that as a second arg to fetchProducts(). Oh and, of course, make this method async.
Up in created, we only need this.loadProducts(null). I'm using null because when we first load, there will be no search term. We could pass this.searchTerm... but I'm going to delete that data in a minute.
This was just a simple refactoring. And... if we reload the page! Yay, our refactoring did not make the site catch on fire. A win!
Back in the editor, the last step is to call this.loadProducts() whenever the search changes... which is exactly when onSearchProducts() is called! Add this.loadProducts() and pass event.term.
Thanks to this, when onSearchProducts() is called, this will start the AJAX call. Later, when it finishes, the products data will get updated and the component will re-render.
And hmm. If you think about it: we don't even need the searchTerm data anymore. I'll look for it: we're setting it here... and initializing it in data. The loadProducts() method doesn't need it because we pass it as an argument.
So celebrate by removing the searchTerm data.
Moment of truth! Let's refresh the page to be safe and... type. Yes! This matches 2 products and disk matches one. In Symfony's web debug toolbar, you can see the AJAX calls. We now have the ability to make our search as powerful as our heart desires.