43.3. Async Computed Properties{Destructured Event Arg}
Destructured Event Arg
Oh, but there's one tiny other thing I want to show you. In onSearchProducts, we get the term with event.term. And... that's actually the only part of the event object that we're using. In the JavaScript world, you'll often see a method like this written as { term }.
This is object destructuring: it grabs the term property from the event argument that's being passed and sets it as a term variable. It allows us to just say term below. We can even document this: the term param is a string. Extra credit if you describe the function above.
After this change... the search still works. But... wow! This is making a lot of AJAX calls! Even if we type really fast, it makes one AJAX call per letter! Let's fix that next by adding debouncing.
Oh, but before we do, it's not hurting anything, but since the created() function does not directly use await anymore, it doesn't need to be async.
Ok, onto the debouncing!