44.1. Debouncing: Data can Hold Anything{Adding the setTimeout()}
The only problem is that we made our search too awesome. When I type... wow! Look at those are AJAX requests - one for every character I type. It's, sort of unnecessarily flooding our API.
This is a common problem with a common solution: debouncing, which is almost as fun as it sounds. With debouncing, instead of sending an AJAX request after every letter, we wait until the user stops typing - maybe 200 milliseconds - and then make the request.
The cool thing is that we're going to be able to add debouncing entirely inside of the search-bar component. This means that our catalog code won't need to change at all; it will instantly be able to take advantage of it.
How can we do that? By delaying the custom search-products event until the user is done typing.

Adding the setTimeout()
Start by adding setTimeout(), passing it an arrow function, then moving the $emit call inside. Set the timeout to 200 milliseconds. Now, instead of emitting the event immediately, it will be slightly delayed.
Easy peasy! Oh, and the arrow function is important: if we used a traditional function, the this variable wouldn't be our Vue instance. Silly JavaScript!
Now, some of you probably realize that this isn't going to quite work yet. If we refresh... and then type really fast. Ah! It still sent four AJAX requests... it just waited 200 milliseconds before making each of them. Whoops!