44.2. Debouncing: Data can Hold Anything{Storing and Clearing the Timeout; Non-Reactive Stuff on Data?}
Storing and Clearing the Timeout
To get debouncing to work, what we really need to do is, onInput(), if a timeout is currently active and waiting to be called, we need to cancel it and restart the timer.
To do that, we need to keep track of the return value of setTimeout(): it's a "timeout id". Then, the next time onInput() is executed we will call clearTimeout() and pass it that id.
None of this is too complex... but I do have one question: where should we store this "timeout id" so that we can reference it later? The easiest thing to do is to store it as data.
Add a new data called searchTimeout set to null.
Then, in the function, say this.searchTimeout = setTimeout(). Now that this is stored, at the top of the function, check if it has a value: if this.searchTimeout, then clearTimeout(this.searchTimeout).
Oh, and to round this all out nicely, once the callback is finally called, we can reset the searchTimeout back to null.
Now, if we type really fast, the second time onInput() is called, it will clear the timeout, and then, below, start a new one.
Let's try it! I'll refresh to be sure then... type super fast. Yes! Just one AJAX call down here to the categories API. That's beautiful!

Non-Reactive Stuff on Data?
For me, the most interesting thing about what we just did is that, in some ways, storing searchTimeout as data seems like an abuse of data. Normally we put something in data because we want it to be reactive: when that value changes, we want any component that uses it to re-render. For searchTimeout... we don't really need that! We just needed a place to stash that value. But... this is fine. The main purpose of data is to store "reactive" data. But if you need a place to store something else, go nuts.
Next, if we have some business logic that we want to re-use between components, where should that live? Let's take our organization up to the next level!