45.2. Business Logic Helpers{Using the Helper in our Component code}
Using the Helper in our Component code
So... we can't just import formatPrice and expect it to magically be available in the template. But we can use it in our JavaScript code, like in our computed property.
Change the template code back to price. Now, in the computed method, use the new helper: return formatPrice(), and pass the same thing we did before: this.product.price.
This time, when we move over... yes! It works perfectly! Let me get rid of the search term and... nice! No errors in the console.
Next: Let's make our search bar a little bit fancier by adding an X icon on the right to clear the search!