46.2. Adding an [x] to our Search Bar{Emit on Clear Search; Check that it all Works! }
Emit on Clear Search
When we think about the search-bar component, the searchTerm data is entirely internal. The only reason we have a searchTerm data at all is just to make our life easier... inside that component. For example, it helps us up here to know whether or not we should hide or show the X button.
But for people that use this component, all they know and care about is that search-bar emits a search-products event. And that is what we're missing in eraseSearchTerm()! We correctly updated the internal data, but we forgot the most important part: emitting the event.
Copy the $emit() statement and use it here. Of course we know searchTerm will always be empty, but this will work fine.

Check that it all Works!
Now when we go over... search for something cool... and clear it... yea! We got it! Making this work was nothing new, but it was a great exercise to think about the internal and external parts of our search-input component.
Next let's do something I'm really excited about! The currentCategoryId is something that we set on page load and then never change. But we've organized our app in such a good way that with a little bit of logic and a new concept called a watcher, we're going to be able to dynamically change the category and have the whole page update. Woh. Let's do this!